Cleaning my courses download a bit
